create carousel pagination component
